Test: Move identifier detection unit tests to CompositeIdFetcherTest
